Fixes #164
This commit fixes #164 by removing JavaFx objects lookup and interacting
directly with MenuItem. The following changes were made:

- previous DahuApp.java was renamed to DahuAppImpl.java and now implements DahuApp interface
- Editor's build.gradle was modified accordingly
- DahuApp.java now contains the DahuApp interface specification
- DahuAppKernel is now initialized with a DahuApp instance
- ContextManager and FileSystem kernel modules were modified accordingly
